# AA-164: Performance fixes for course home