--- a/lms/templates/courseware/info.html
+++ b/lms/templates/courseware/info.html
@@ -9,7 +9,7 @@ from pytz import timezone, utc
 from django.urls import reverse
 from django.utils.translation import ugettext as _
 
-from lms.djangoapps.courseware.courses import get_course_info_section, get_course_date_blocks
+from lms.djangoapps.courseware.courses import get_course_info_section
 from openedx.core.djangoapps.self_paced.models import SelfPacedConfiguration
 from openedx.core.djangolib.markup import HTML, Text
 %>