Local course home page load averages
```
PLS Disabled w/o Changes: 3.52s
PLS Enabled w/o Changes: 6.69s
PLS Disabled w/ Changes: 2.88s
PLS Enabled w/ Changes: 3.73s
```